requirements.txt
requests==2.32.5
pandas==2.3.3
Faker==37.12.0
numpy==2.0.2
apache-airflow-providers-google==16.0.0
dbt-bigquery==1.10.2